Refactor: Move onCodeLens to its own file
